fix: abstract logic
